# docs/05-api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Klick-Pro (servio) HTTP API
  version: 0.0.0-reverse-engineered
  description: >-
    Reverse-engineered from source at commit cd8f4fb (verified 2026-09-16). Source of truth is the implementation; every
    operation carries x-source file:line.


    Conventions and caveats:

    - Responses are raw NextResponse.json; errors are mostly {"error": string}. The apiSuccess/apiError envelope in
    src/lib/api-response.ts is not used.

    - Every POST/PUT/PATCH/DELETE under /api/* must carry an Origin header equal to the app origin or APP_URL, otherwise
    proxy.ts returns 403 {"error":"Request origin is not allowed."}. This applies to webhooks and non-browser clients
    too. Under the custom server only the exact APP_URL origin passes in practice; a genuine same-origin request to a
    different host (e.g. 127.0.0.1 when APP_URL is localhost) also gets 403. [VALIDATED 2026-09-17 · V-20]

    - /api/v1/:path* is rewritten to /api/:path* (next.config.ts, afterFiles). Physical routes /api/v1/messages and
    /api/v1/professionals take precedence and have no /api/* twin. /api/v1/v1/* is not rewritten twice and returns
    404. [VALIDATED 2026-09-17 · V-13]

    - Schemas marked [NEEDS VALIDATION] could not be fully established from code; no properties were invented.

    - Realtime Socket.IO events (path /api/realtime) are documented in docs/05-api/api-specification.md, not here.


    Human-readable reference: docs/05-api/api-specification.md
servers:
...
paths:
...
  /api/admin/users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
    get:
      tags:
        - Admin
      operationId: adminGetUser
      summary: User detail and stats
      x-source: app/api/admin/users/[id]/route.ts:59-155
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User detail.
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                  - stats
                properties:
                  user:
                    $ref: "#/components/schemas/PrismaModel"
                  stats:
                    type: object
                    properties:
                      jobsPosted:
                        type: integer
                      proposals:
                        type: integer
                      projects:
                        type: integer
                      completedPayments:
                        type: integer
                      money:
                        type: number
                      services:
                        type: integer
        "400":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "403":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "500":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    patch:
      tags:
        - Admin
      operationId: adminUpdateUserActive
      summary: Activate/deactivate user (deactivation revokes sessions)
      x-source: app/api/admin/users/[id]/route.ts:6-34
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - isActive
              properties:
                isActive:
                  type: boolean
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      isActive:
                        type: boolean
        "400":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "403":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "500":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    delete:
      tags:
        - Admin
      operationId: adminDeleteUser
      summary: Hard-delete user
      x-source: app/api/admin/users/[id]/route.ts:36-57
      security:
        - cookieAuth: []
      responses:
        "200":
          description: "{ok: true}"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
        "400":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "403":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "500":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
  /api/dashboard:
    get:
      tags:
        - Dashboard
      operationId: clientDashboard
      summary: Client dashboard aggregate
      x-source: app/api/dashboard/route.ts:5-103
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Dashboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDashboard"
        "401":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "403":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "500":
          description: "Error with a human-readable message: {error: string}."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
  /api/professional/earnings/export:
    post:
      tags:
        - Professional
      summary: PDF of earnings (ProjectTransaction) and payouts (ProjectWithdrawal)
      operationId: exportProfessionalEarnings
      security:
        - cookieAuth: []
      description: Source app/api/professional/earnings/export/route.ts:54. scope=selected applies ids to both tables.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
      responses:
        "200":
          description: PDF attachment
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid export request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not an active PROFESSIONAL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Report could not be generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: servio_session
      description: >-
        HS256 JWT {userId, role, sessionId} issued by src/lib/auth.ts createSession; validated against the sessions
        table on every request. Same cookie for clients, professionals and admins — role is inside the token.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Same session JWT sent as Authorization: Bearer. Accepted only by the few handlers noted per operation."
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    OkTrue:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
          const: true
    ReportRequest:
      type: object
      required:
        - scope
      properties:
        scope:
          type: string
          enum:
            - all
            - selected
        ids:
          type: array
          items:
            type: integer
            minimum: 1
        pageSize:
          type: string
          enum:
            - A4
            - LETTER
        orientation:
          type: string
          enum:
            - portrait
            - landscape
      x-source: src/lib/reports/pdf/request.ts:4-9
...
    ErrorMessage:
      type: object
      required:
        - error
      properties:
        error:
          type: string
      x-source: e.g. app/api/auth/[action]/route.ts:44
...
    PrismaModel:
      type: object
      description: >-
        Serialized Prisma model row (optionally with includes). Field list follows prisma/schema.prisma; see
        docs/04-database/data-dictionary.md. [NEEDS VALIDATION] per-endpoint field selection is described in
        api-specification.md.
...
    ClientDashboard:
      type: object
      required:
        - user
        - jobs
        - projectSummary
        - proposals
        - hireRequests
        - notifications
        - spent
      properties:
        user:
          type: object
          properties:
            firstName:
              type: string
            averageRating:
              type:
                - number
                - "null"
            phoneVerifiedAt:
              type:
                - string
                - "null"
              format: date-time
        jobs:
          type: array
          maxItems: 4
          items:
            $ref: "#/components/schemas/PrismaModel"
          description: ClientJob + _count.favoriteJobs; status may be RUNNING
        projectSummary:
          type: object
          properties:
            total:
              type: integer
            open:
              type: integer
            running:
              type: integer
            drafts:
              type: integer
        proposals:
          type: array
          maxItems: 4
          items:
            $ref: "#/components/schemas/PrismaModel"
          description: ProjectRequest (PROFESSIONAL_PROPOSAL) + professionalName
        hireRequests:
          type: array
          maxItems: 4
          items:
            $ref: "#/components/schemas/PrismaModel"
          description: ProjectRequest (CLIENT_HIRE) + professionalName
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/PrismaModel"
          description: UserNotification (unbounded)
        spent:
          type: number
      x-source: app/api/dashboard/route.ts:80-99
...